cleanUp: delete duplicate functions and variables
