+++ config/auth.js
@@ -14,7 +14,6 @@ let generateJWTToken = (user) => {
     });
 }
 
-/* login. */
 module.exports = (router) => {
     router.post('/login', (req, res) => {
         passport.authenticate('local', { session: false }, (error, user, info) => {
